requirements.txt
yt-dlp
opencv-python
Pillow
requests
numpy
torch
transformers
